FIX wrong marshal of void to IntPtr and bump version to 3.7.1.

+++ src/Alimer.Bindings.SDL/SDL.Gamepad.cs
@@ -102,7 +102,7 @@ unsafe partial class SDL3
     {
         fixed (void* dataPtr = &data)
         {
-            return SDL_SendGamepadEffect(gamepad, dataPtr, size);
+            return SDL_SendGamepadEffect(gamepad, (nint)dataPtr, size);
         }
     }
